chore: rename api directory to network
