+++ libs/coin-polkadot/src/bridge/prepareTransaction.ts
@@ -1,7 +1,7 @@
 import type { PolkadotAccount, Transaction } from "../types";
 import getEstimatedFees from "./getFeesForTransaction";
 import { loadPolkadotCrypto } from "./polkadot-crypto";
-import { PolkadotAPI } from "../api";
+import { PolkadotAPI } from "../network";
 import BigNumber from "bignumber.js";
 
 const sameFees = (a: BigNumber, b?: BigNumber | null) => (!a || !b ? a === b : a.eq(b));
