+++ libs/coin-polkadot/src/bridge/signOperation.ts
@@ -20,7 +20,7 @@ import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { buildTransaction } from "./buildTransaction";
 import { calculateAmount, getNonce, isFirstBond } from "./logic";
-import { PolkadotAPI } from "../api";
+import { PolkadotAPI } from "../network";
 import { PolkadotAddress, PolkadotSignature, PolkadotSigner } from "../types";
 
 const MODE_TO_TYPE = {
